+++ R/ncToData.R
@@ -60,6 +60,10 @@ ncToData <- function(data, nc, var=NULL, buffer = c(0,0,0), FUN = c(mean),
     nc <- romsCheck(nc)
     oldNames <- colnames(data)
     colnames(data) <- standardCoordNames(oldNames)
+    if(all(is.na(data$Longitude)) ||
+       all(is.na(data$Latitude))) {
+        return(data)
+    }
     # this finds the name of the longitude coordinate from my list that matches lon/long/whatever to Longitude
     nc180 <- ncIs180(nc)
     data180 <- dataIs180(data)
@@ -92,7 +96,7 @@ ncToData <- function(data, nc, var=NULL, buffer = c(0,0,0), FUN = c(mean),
     usedDim <- unique(unlist(sapply(nc$var, function(x) x$dimids+1)))
     usedDim <- usedDim[!is.na(usedDim)]
     names(nc$dim)[usedDim] <- standardCoordNames(names(nc$dim)[usedDim])
-        
+
     if('Depth' %in% names(nc$dim)) {
         matchDims <- c('matchLong', 'matchLat', 'matchTime', 'matchDepth')
     } else {
@@ -164,7 +168,7 @@ ncToData <- function(data, nc, var=NULL, buffer = c(0,0,0), FUN = c(mean),
         }
         if(!is.null(data$matchTime_mean) &&
            !(all(is.na(data$matchTime_mean))) &&
-           max(abs(data$matchTime_mean)) > 365) {
+           max(abs(data$matchTime_mean), na.rm=TRUE) > 365) {
             data$matchTime_mean <- as.POSIXct(data$matchTime_mean, origin = '1970-01-01 00:00:00', tz='UTC')
         }
     }
@@ -203,8 +207,25 @@ getVarData <- function(data, nc, var, buffer, depth=NULL, verbose=TRUE) {
             zVals <- nc$dim$Depth$vals[zIx$ix]
         }
     }
-    result <- vector('list', length = length(var) + 3)
-    names(result) <- c(var, 'matchLong', 'matchLat', 'matchTime')
+    result <- vector('list', length = length(var) + 3 + hasZ)
+    outNames <- c(var, 'matchLong', 'matchLat', 'matchTime')
+    if(hasZ) {
+        outNames <- c(outNames, 'matchDepth')
+    }
+    names(result) <- outNames
+    checkNa <- c(xIx$start, xIx$count, yIx$start, yIx$count)
+    if(hasT) {
+        checkNa <- c(checkNa, tIx$start, tIx$count)
+    }
+    if(hasZ) {
+        checkNa <- c(checkNa, zIx$start, zIx$count)
+    }
+    if(any(is.na(checkNa))) {
+        for(r in seq_along(result)) {
+            result[[r]] <- NA
+        }
+        return(result)
+    }
     for(v in var) {
         thisVar <- nc[['var']][[v]]
         thisDimId <- thisVar$dimids+1
@@ -230,7 +251,7 @@ getVarData <- function(data, nc, var, buffer, depth=NULL, verbose=TRUE) {
                    'UTC' = {
                        if(multiT) {
                            tIx <- dimToIx(data$UTC, nc$dim[[thisDimId[d]]], buffer[3], verbose)
-                           tVals <- ncTimeToPosix(nc$dim[[thisDimId[d]]]$vals[tIx$ix], 
+                           tVals <- ncTimeToPosix(nc$dim[[thisDimId[d]]]$vals[tIx$ix],
                                               units = nc$dim[[thisDimId[d]]]$units)
                        }
                        start <- c(start, tIx$start)